requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.1
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2